Message section height
Message section height separated for both desktop and mobile view. for mobile view it is set to fit-content so that blank spaces doesnt appear in mobile view and for desktop view height is increased to 80vh to perfectly fit hrithik sir message in message box as its overflowing from it.